% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_to_file}
\alias{pack_to_file}
\title{Pack a CWL workflow and output a file}
\usage{
pack_to_file(input, output)
}
\arguments{
\item{input}{Path to the input CWL file}

\item{output}{Path to the output CWL file}
}
\description{
Pack a CWL workflow and output a file
}
\examples{
input <- system.file("example/remote-cwl/wf1.cwl", package = "sbpack")
output <- tempfile(fileext = ".cwl")
pack_to_file(input, output)
cat(readLines(output), sep = "\n")

input <- "https://raw.githubusercontent.com/rabix/sbpack/master/tests/wf2.cwl"
output <- tempfile(fileext = ".cwl")
pack_to_file(input, output)
cat(readLines(output), sep = "\n")
}
