% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_to_sb}
\alias{pack_to_sb}
\title{Pack and upload CWL apps to any Seven Bridges powered platform}
\usage{
pack_to_sb(input, profile, id)
}
\arguments{
\item{input}{Path to the input CWL file. Can be a remote file.}

\item{profile}{Platform profile name defined in \code{~/.sevenbridges/credentials}.}

\item{id}{Destination app id in the form of \code{{user}/{project}/{app_id}}.
This will install (or update) \code{app_id} located in \code{project}
of \code{user} on the platform.}
}
\description{
Pack and upload CWL apps to any Seven Bridges powered platform
}
\examples{
\dontrun{
input <- system.file("example/remote-cwl/wf1.cwl", package = "sbpack")
pack_to_sb(input, "aws-us", "username/project/workflow")}
}
