% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{py_pack_file}
\alias{py_pack_file}
\title{Global reference to py_pack_file}
\usage{
py_pack_file(...)
}
\arguments{
\item{...}{not used}
}
\value{
py_pack_file function
}
\description{
Global reference to py_pack_file which will be initialized in \code{.onLoad}.
}
