% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{py_pack_sb}
\alias{py_pack_sb}
\title{Global reference to py_pack_sb}
\usage{
py_pack_sb(...)
}
\arguments{
\item{...}{not used}
}
\value{
py_pack_sb function
}
\description{
Global reference to py_pack_sb which will be initialized in \code{.onLoad}.
}
