% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{pysbpack}
\alias{pysbpack}
\title{Global reference to sbpack}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
pysbpack
}
\value{
sbpack reference object
}
\description{
Global reference to sbpack which will be initialized in \code{.onLoad}.
}
\keyword{datasets}
