% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sbpack-package}
\alias{sbpack}
\alias{sbpack-package}
\title{sbpack: Pack and Upload CWL Apps to Seven Bridges Powered Platforms}
\description{
An R wrapper for the Python package 'sbpack'.
    Packs and uploads CWL apps to Seven Bridges powered platforms.
    Resolves linked processes, schemadefs, $include, and $import.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/sbpack-r/}
  \item \url{https://github.com/nanxstats/sbpack-r}
  \item Report bugs at \url{https://github.com/nanxstats/sbpack-r/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao@sevenbridges.com} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item Kaushik Ghose \email{kaushik.ghose@sevenbridges.com} (\href{https://orcid.org/0000-0003-2933-1260}{ORCID})
}

Other contributors:
\itemize{
  \item Seven Bridges Genomics [copyright holder, funder]
}

}
\keyword{internal}
